Moviendo modelos a dominio
